Added responsive UI features
The whole app now feels good to use on the phone. There's a slight issue
with the AddTransaction page where the textboxes appear to long in the
desktop version and there is a slight padding issue with the walletView
page. But other than that, the whole app looks crisp and responsive.

+++ client/src/components/walletView/TransView.js
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 import { useHistory } from 'react-router-dom';
-import Backend from '../apis/Backend';
-import { WalletContext } from '../context/WalletContext';
+import Backend from '../../apis/Backend';
+import { WalletContext } from '../../context/WalletContext';
 
 const typeColor = {
     Expense: "bg-danger",
@@ -16,7 +16,10 @@ const TransView = ({ transactions, setTransactions }) => {
     const { selectedWallet, setSelectedWallet } = useContext(WalletContext);
     
     const handleDelete = async (trans) => {
-        if (window.confirm(`Are you sure you want to delete transaction: ${trans.details}`)){
+        if (trans.details === 'Starting Balance'){
+            return null;
+        }
+        else if (window.confirm(`Are you sure you want to delete transaction: ${trans.details}`)){
             try {
                 const response = await Backend.delete(`/transactions/${trans.id}`);
                 setTransactions(transactions.filter(transaction=>trans.id !== transaction.id));
@@ -28,26 +31,26 @@ const TransView = ({ transactions, setTransactions }) => {
     }
 
     return (
-        <table className="table table-hover">
+        <table className="table">
             <thead>
                 <tr>
                 <th scope="col">Date</th>
-                <th scope="col">Type</th>
+                {/*<th scope="col">Type</th>*/}
                 <th scope="col">Details</th>
                 <th scope="col">Category</th>
-                <th scope="col">Amount</th>
-                <th scope="col">Action</th>
+                <th scope="col">Amt</th>
+                {/*<th scope="col">Action</th>*/}
                 </tr>
             </thead>
             <tbody>
                 {transactions && transactions.map(transaction => {
-                    return (<tr key={transaction.id} className={typeColor[transaction.type]}>
+                    return (<tr onClick={()=>handleDelete(transaction)} key={transaction.id} className={typeColor[transaction.type]}>
                         <td>{transaction.date}</td>
-                        <td>{transaction.type}</td>
+                        {/*<td>{transaction.type}</td>*/}
                         <td><strong>{transaction.details}</strong></td>
                         <td>{transaction.category_name}</td>
                         <td>{transaction.amount < 0 ? "("+-1*transaction.amount+")" : transaction.amount}</td>
-                        <td>{transaction.type !== "Starting" ? <p style={{ cursor: "pointer" }} onClick={()=>handleDelete(transaction)}><strong>Remove</strong></p>:null}</td>
+                        {/*<td>{transaction.type !== "Starting" ? <p style={{ cursor: "pointer" }} onClick={()=>handleDelete(transaction)}><strong>Remove</strong></p>:null}</td>*/}
                     </tr>)
                 })}
             </tbody>
